+++ src/cpp/flann/algorithms/kdtree_index.h
@@ -152,7 +152,7 @@ public:
         else {
             for (size_t i=old_size;i<size_;++i) {
                 for (int j = 0; j < trees_; j++) {
-                    addPointToTree(tree_roots_[j], i);
+                    addPointToTree(tree_roots_[j], (int)i);
                 }
             }
         }        
@@ -691,7 +691,7 @@ private:
                 right->divfeat = ind;
                 right->point = point;
             }
-            node->divfeat = div_feat;
+            node->divfeat = (int)div_feat;
             node->divval = (point[div_feat]+leaf_point[div_feat])/2;
             node->child1 = left;
             node->child2 = right;
